+++ pulldown-cmark-mdcat/tests/wrapping.rs
@@ -14,17 +14,19 @@ use pulldown_cmark::{Options, Parser};
 use syntect::parsing::SyntaxSet;
 
 use anyhow::{Context, Result};
-use mdcat::{resources::NoopResourceHandler, terminal::TerminalProgram, Environment, Theme};
+use pulldown_cmark_mdcat::resources::NoopResourceHandler;
+use pulldown_cmark_mdcat::terminal::{TerminalProgram, TerminalSize};
+use pulldown_cmark_mdcat::{Environment, Settings, Theme};
 
 static SYNTAX_SET: Lazy<SyntaxSet> = Lazy::new(SyntaxSet::load_defaults_newlines);
-static SETTINGS_ANSI_ONLY: Lazy<mdcat::Settings> = Lazy::new(|| mdcat::Settings {
+static SETTINGS_ANSI_ONLY: Lazy<Settings> = Lazy::new(|| Settings {
     terminal_capabilities: TerminalProgram::Ansi.capabilities(),
-    terminal_size: mdcat::terminal::TerminalSize::default(),
+    terminal_size: TerminalSize::default(),
     theme: Theme::default(),
     syntax_set: &SYNTAX_SET,
 });
 
-fn render_to_string<S: AsRef<str>>(markdown: S, settings: &mdcat::Settings) -> Result<String> {
+fn render_to_string<S: AsRef<str>>(markdown: S, settings: &Settings) -> Result<String> {
     let parser = Parser::new_ext(
         markdown.as_ref(),
         Options::ENABLE_TASKLISTS | Options::ENABLE_STRIKETHROUGH,
@@ -34,7 +36,7 @@ fn render_to_string<S: AsRef<str>>(markdown: S, settings: &mdcat::Settings) -> R
         hostname: "HOSTNAME".to_string(),
         ..Environment::for_local_directory(&std::env::current_dir()?)?
     };
-    mdcat::push_tty(settings, &env, &NoopResourceHandler, &mut sink, parser)?;
+    pulldown_cmark_mdcat::push_tty(settings, &env, &NoopResourceHandler, &mut sink, parser)?;
     String::from_utf8(sink).with_context(|| "Failed to convert rendered result to string")
 }